feat(dashboard): brand wordmark on the clock bar + fold Settings into the ⋯ menu (calvin-800)
- BarLogo renders the CAL·VIN wordmark (matching the Settings shell topbar)
  instead of the faint arc-C glyph; adds a vertical-rl variant for side bars.
- Move the standalone gear Settings button out of BarActionCluster and into
  the AdminOverflow (⋯) menu as its first item, decluttering the bar.
