+++ frontend/src/components/BarActionCluster.vue
@@ -14,16 +14,6 @@
     </div>
 
     <template v-if="configStore.shouldShowUI">
-      <button
-        class="bar-btn"
-        :title="compact ? 'Settings' : undefined"
-        aria-label="Open settings"
-        @click="goToSettings"
-      >
-        <span class="bar-btn-icon">⚙️</span>
-        <span v-if="!compact" class="bar-btn-label">Settings</span>
-      </button>
-
       <AdminOverflow />
     </template>
   </div>
@@ -32,9 +22,7 @@
 <script setup>
 import { ref, computed, onMounted, onUnmounted } from "vue";
 import axios from "axios";
-import { useRouter } from "vue-router";
 import { useConfigStore } from "../stores/config";
-import { useModeStore } from "../stores/mode";
 import ConnectionIndicator from "./ConnectionIndicator.vue";
 import AdminOverflow from "./dashboard/AdminOverflow.vue";
 
@@ -48,8 +36,6 @@ defineProps({
 const visible = computed(() => true);
 
 const configStore = useConfigStore();
-const modeStore = useModeStore();
-const router = useRouter();
 
 const status = ref("checking...");
 let healthInterval = null;
@@ -73,11 +59,6 @@ const checkHealth = async () => {
       status.value = "error";
     }
   }
-};
-
-const goToSettings = () => {
-  modeStore.setMode(modeStore.MODES.SETTINGS);
-  router.push("/settings");
 };
 
 onMounted(() => {
@@ -149,40 +130,4 @@ onUnmounted(() => {
   }
 }
 
-.bar-btn {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.4rem;
-  background: var(--bg-2);
-  color: var(--ink);
-  border: 1px solid var(--line);
-  border-radius: 4px;
-  padding: 0.35rem 0.6rem;
-  font-size: 0.85rem;
-  cursor: pointer;
-  transition:
-    background 0.2s,
-    border-color 0.2s;
-  white-space: nowrap;
-}
-
-.bar-action-cluster.compact .bar-btn {
-  padding: 0.4rem;
-  width: 100%;
-  justify-content: center;
-}
-
-.bar-btn:hover {
-  background: var(--bg-2);
-  border-color: var(--ink-2);
-}
-
-.bar-btn-icon {
-  font-size: 1rem;
-  line-height: 1;
-}
-
-.bar-btn-label {
-  font-size: 0.85rem;
-}
 </style>
